refactor: adding the mongoose schemas
