--- back/fillter/src/entities/users/eatery.js
@@ -1,19 +1,0 @@
-class Eatery {
-    constructor(name, address, category) {
-        this.name = name;
-        this.address = address;
-        this.category = category;
-    }
-    getName(){
-        return this.name;
-    }
-    getAddress(){
-        return this.address;
-    }
-    getCategory(){
-        return this.category;
-    }
-
-}
-
-export default Eatery;
